exported Domain, Lollipop, and Sequence specs
