+++ packages/react-mutation-mapper/src/Lollipop.tsx
@@ -1,7 +1,8 @@
 import * as React from 'react';
 import {observer} from "mobx-react";
 import {computed, observable} from "mobx";
-import {LollipopSpec} from "./LollipopPlotNoTooltip";
+
+import LollipopSpec from "./model/LollipopSpec";
 
 type LollipopProps = {
     x:number;
