+++ packages/react-mutation-mapper/src/index.tsx
@@ -3,3 +3,7 @@ export {default as LollipopPlot} from "./LollipopPlot";
 export {default as Domain} from "./Domain";
 export {default as Lollipop} from "./Lollipop";
 export {default as Sequence} from "./LollipopPlot";
+export {default as DomainSpec} from "./model/DomainSpec";
+export {default as LollipopSpec} from "./model/LollipopSpec";
+export {default as SequenceSpec} from "./model/SequenceSpec";
+
