+++ packages/react-mutation-mapper/src/model/DataStore.ts
@@ -1,10 +1,12 @@
 export interface DataStore {
-  setPositionSelected: (position:number, newVal:boolean) => void;
-  setPositionHighlighted: (position:number, newVal:boolean) => void;
-  clearSelectedPositions: () => void;
-  clearHighlightedPositions: () => void;
-  clearDataSelectFilter: () => void;
-  clearDataHighlightFilter: () => void;
-  isPositionSelected: (position:number) => boolean;
-  isPositionHighlighted: (position:number) => boolean;
+    setPositionSelected: (position:number, newVal:boolean) => void;
+    setPositionHighlighted: (position:number, newVal:boolean) => void;
+    clearSelectedPositions: () => void;
+    clearHighlightedPositions: () => void;
+    clearDataSelectFilter: () => void;
+    clearDataHighlightFilter: () => void;
+    isPositionSelected: (position:number) => boolean;
+    isPositionHighlighted: (position:number) => boolean;
 }
+
+export default DataStore;
