+++ packages/react-mutation-mapper/src/model/DomainSpec.ts
@@ -1,0 +1,10 @@
+export type DomainSpec = {
+    startCodon:number;
+    endCodon:number;
+    color:string;
+    label?:string;
+    labelColor?:string;
+    tooltip?:JSX.Element;
+};
+
+export default DomainSpec;
