+++ packages/react-mutation-mapper/src/model/SequenceSpec.ts
@@ -1,0 +1,5 @@
+export type SequenceSpec = {
+    tooltip?:JSX.Element;
+};
+
+export default SequenceSpec;
